requirements.txt
Django==5.0.7
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3